initial dash data

+++ backend/app/Http/Controllers/API/Admin/CourierController.php
@@ -93,7 +93,7 @@ class CourierController extends Controller
         if($image != '')
         {
           $image_path = "images/courier/".$image_name;
-          if(File::exists($image_path)) {
+          if(($image_name != 'no-image.png') && (File::exists($image_path))) {
             File::delete($image_path);
         }
         $image_name = rand() . '.' . $image->getClientOriginalExtension();
